move WalletScreen to its own directory
