+++ src/screens/WalletScreen/WalletScreen.js
@@ -3,9 +3,9 @@ import { compact, isEmpty, keys } from 'lodash';
 import React, { useEffect, useState } from 'react';
 import { InteractionManager } from 'react-native';
 import { useDispatch, useSelector } from 'react-redux';
-import { OpacityToggler } from '../components/animations';
-import { AssetList } from '../components/asset-list';
-import { Page } from '../components/layout';
+import { OpacityToggler } from '../../components/animations';
+import { AssetList } from '../../components/asset-list';
+import { Page } from '../../components/layout';
 import { Network } from '@/helpers';
 import { useRemoveFirst } from '@/navigation/useRemoveFirst';
 import { settingsUpdateNetwork } from '@/redux/settings';
@@ -32,7 +32,7 @@ import {
 import { useNavigation } from '@/navigation';
 import { updateRefetchSavings } from '@/redux/data';
 import { emitChartsRequest, emitPortfolioRequest } from '@/redux/explorer';
-import Routes from '@/navigation/routesNames';
+import Routes from '@rainbow-me/routes';
 import styled from '@/styled-thing';
 import { position } from '@/styles';
 import { Toast, ToastPositionContainer } from '@/components/toasts';
